requirements.txt
PyPDF2
pdf2image
PyMuPDF
reportlab
openpyxl
pandas
Pillow
pytesseract
opencv-python
clipboard
tqdm
pyinstaller
